+++ runtime/common/Cargo.toml
@@ -8,8 +8,10 @@ version.workspace = true
 
 [dependencies]
 # crates.io
-codec    = { workspace = true, package = "parity-scale-codec" }
-smallvec = { version = "1.10" }
+codec      = { workspace = true, package = "parity-scale-codec" }
+scale-info = { workspace = true }
+smallvec   = { version = "1.10" }
+
 
 # darwinia
 darwinia-precompile-assets = { workspace = true }
@@ -44,6 +46,7 @@ default = ["std"]
 std = [
 	# crates.io
 	"codec/std",
+	"scale-info/std",
 
 	# darwinia
 	"darwinia-precompile-assets/std",
