+++ runtime/pangolin/src/pallets/orml_xtokens.rs
@@ -31,38 +31,7 @@ pub enum CurrencyId {
 	SelfReserve,
 	// Assets representing other chains native tokens
 	ForeignAsset(crate::AssetId),
-	// Our local assets. We don't need this.
-	// LocalAssetReserve(AssetId),
-	// Erc20 token
-	// Erc20 { contract_address: sp_core::H160 },
 }
-
-// TODO erc20 token
-// impl xcm_primitives::AccountIdToCurrencyId<AccountId, CurrencyId> for Runtime {
-// 	fn account_to_currency_id(account: AccountId) -> Option<CurrencyId> {
-// 		Some(match account {
-// 			// the self-reserve currency is identified by the pallet-balances address
-// 			a if a == sp_core::H160::from_low_u64_be(2050).into() => CurrencyId::SelfReserve,
-// 			// the rest of the currencies, by their corresponding erc20 address
-// 			_ => match Runtime::account_to_asset_id(account) {
-// 				// We distinguish by prefix, and depending on it we create either
-// 				// Foreign or Local
-// 				Some((prefix, asset_id)) => {
-// 					if prefix == FOREIGN_ASSET_PRECOMPILE_ADDRESS_PREFIX.to_vec() {
-// 						CurrencyId::ForeignAsset(asset_id)
-// 					} else {
-// 						CurrencyId::LocalAssetReserve(asset_id)
-// 					}
-// 				}
-// 				// If no known prefix is identified, we consider that it's a "real" erc20 token
-// 				// (i.e. managed by a real smart contract)
-// 				None => CurrencyId::Erc20 {
-// 					contract_address: account.into(),
-// 				},
-// 			},
-// 		})
-// 	}
-// }
 
 // How to convert from CurrencyId to MultiLocation
 pub struct CurrencyIdtoMultiLocation<AssetXConverter>(sp_std::marker::PhantomData<AssetXConverter>);
@@ -78,22 +47,6 @@ where
 				Some(multi)
 			},
 			CurrencyId::ForeignAsset(asset) => AssetXConverter::reverse_ref(asset).ok(),
-			// No transactor matches this yet, so even if we have this enum variant the transfer
-			// will fail CurrencyId::LocalAssetReserve(asset) => {
-			// 	let mut location = LocalAssetsPalletLocation::get();
-			// 	location.push_interior(Junction::GeneralIndex(asset)).ok();
-			// 	Some(location)
-			// }
-			// CurrencyId::Erc20 { contract_address } => {
-			// 	let mut location = Erc20XcmBridgePalletLocation::get();
-			// 	location
-			// 		.push_interior(Junction::AccountKey20 {
-			// 			key: contract_address.0,
-			// 			network: None,
-			// 		})
-			// 		.ok();
-			// 	Some(location)
-			// }
 		}
 	}
 }
@@ -120,11 +73,7 @@ impl orml_xtokens::Config for Runtime {
 	type BaseXcmWeight = BaseXcmWeight;
 	type CurrencyId = CurrencyId;
 	type CurrencyIdConvert = CurrencyIdtoMultiLocation<
-		xcm_primitives::AsAssetType<
-			crate::AssetId,
-			pallets::asset_manager::AssetType,
-			AssetManager,
-		>,
+		xcm_primitives::AsAssetType<crate::AssetId, xcm_configs::AssetType, AssetManager>,
 	>;
 	type MaxAssetsForTransfer = MaxAssetsForTransfer;
 	// We don't have this case: fee_reserve != non_fee_reserve
