+++ runtime/pangolin/src/pallets/polkadot_xcm.rs
@@ -39,7 +39,7 @@ pub type LocalAssetTransactor = xcm_builder::CurrencyAdapter<
 
 // The non-reserve fungible transactor type
 // It will use pallet-assets, and the Id will be matched against AsAssetType
-pub type ForeignFungiblesTransactor = xcm_configs::LimitFungiblesAdapter<
+pub type ForeignFungiblesTransactor = xcm_builder::FungiblesAdapter<
 	// Use this fungibles implementation:
 	Assets,
 	// Use this currency when it is a fungible asset matching the given location or name:
@@ -47,11 +47,7 @@ pub type ForeignFungiblesTransactor = xcm_configs::LimitFungiblesAdapter<
 		xcm_builder::ConvertedConcreteId<
 			crate::AssetId,
 			Balance,
-			xcm_primitives::AsAssetType<
-				crate::AssetId,
-				pallets::asset_manager::AssetType,
-				AssetManager,
-			>,
+			xcm_primitives::AsAssetType<crate::AssetId, xcm_configs::AssetType, AssetManager>,
 			xcm_executor::traits::JustTry,
 		>,
 	),
@@ -205,11 +201,7 @@ pub type XcmFeesToAccount = xcm_primitives::XcmFeesToAccount<
 		xcm_builder::ConvertedConcreteId<
 			crate::AssetId,
 			Balance,
-			xcm_primitives::AsAssetType<
-				crate::AssetId,
-				pallets::asset_manager::AssetType,
-				AssetManager,
-			>,
+			xcm_primitives::AsAssetType<crate::AssetId, xcm_configs::AssetType, AssetManager>,
 			xcm_executor::traits::JustTry,
 		>,
 	),
@@ -254,11 +246,7 @@ impl xcm_executor::Config for XcmExecutorConfig {
 			DealWithFees<Runtime>,
 			ToTreasury,
 		>,
-		xcm_primitives::FirstAssetTrader<
-			pallets::asset_manager::AssetType,
-			AssetManager,
-			XcmFeesToAccount,
-		>,
+		xcm_primitives::FirstAssetTrader<xcm_configs::AssetType, AssetManager, XcmFeesToAccount>,
 	);
 	type UniversalAliases = frame_support::traits::Nothing;
 	// Teleporting is disabled.
